% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cmatern}
\alias{cmatern}
\title{The Matern covariance function as paramterized by \href{https://en.wikipedia.org/wiki/Mat\%C3\%A9rn_covariance_function}{wikipedia}.}
\usage{
cmatern(t, sig, rho, nu)
}
\arguments{
\item{t}{the distance/argument}

\item{sig}{the marginal standard deviation}

\item{rho}{the range parameter}

\item{nu}{the smoothness parameter}
}
\value{
the Matern covariance function with parameters (sig, rho, nu)
evaluated at distance t.
}
\description{
The Matern covariance function as paramterized by \href{https://en.wikipedia.org/wiki/Mat\%C3\%A9rn_covariance_function}{wikipedia}.
}
