% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cmatern_dnu}
\alias{cmatern_dnu}
\title{The derivative of the Matern covariance function with respect to nu.}
\usage{
cmatern_dnu(t, sig, rho, nu)
}
\arguments{
\item{t}{the distance/argument}

\item{sig}{the marginal standard deviation}

\item{rho}{the range parameter}

\item{nu}{the smoothness parameter}
}
\value{
the derivative of the Matern covariance function with respect to
nu with parameters (sig, rho, nu) ' evaluated at distance t.
}
\description{
The derivative of the Matern covariance function with respect to nu.
}
